// frontend/src/features/_utils_/gql/types.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  AccessibilityEnum: { input: any; output: any; }
  /**
   * The `BigInt` scalar type represents non-fractional whole numeric values.
   * `BigInt` is not constrained to 32-bit like the `Int` type and thus is a less
   * compatible type.
   */
  BigInt: { input: any; output: any; }
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: { input: any; output: any; }
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: { input: any; output: any; }
  /** The `Decimal` scalar type represents a python Decimal. */
  Decimal: { input: any; output: any; }
  FinancingEnum: { input: any; output: any; }
  HydrophoneDirectivityEnum: { input: any; output: any; }
  /** Django Primary key */
  PK: { input: any; output: any; }
  RoleEnum: { input: any; output: any; }
  SignalPluralityEnum: { input: any; output: any; }
  SignalShapeEnum: { input: any; output: any; }
  StatusEnum: { input: any; output: any; }
  TypeEnum: { input: any; output: any; }
};
// ...
export type AcousticFeaturesNode = Node & {
  __typename?: 'AcousticFeaturesNode';
  annotation?: Maybe<AnnotationNode>;
  /** [Hz] Frequency at the end of the signal */
  endFrequency?: Maybe<Scalars['Float']['output']>;
  /** If the signal has harmonics */
  hasHarmonics?: Maybe<Scalars['Boolean']['output']>;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  pk: Scalars['PK']['output'];
  /** Number of relative maximum frequency in the signal */
  relativeMaxFrequencyCount?: Maybe<Scalars['Int']['output']>;
  /** Number of relative minimum frequency in the signal */
  relativeMinFrequencyCount?: Maybe<Scalars['Int']['output']>;
  /** [Hz] Frequency at the beginning of the signal */
  startFrequency?: Maybe<Scalars['Float']['output']>;
  /** Number of steps (flat segment) in the signal */
  stepsCount?: Maybe<Scalars['Int']['output']>;
  trend?: Maybe<SignalTrendType>;
};
// ...
export type AnnotationCampaignNodeConnection = {
  __typename?: 'AnnotationCampaignNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<AnnotationCampaignNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type AnnotationCommentNodeConnection = {
  __typename?: 'AnnotationCommentNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<AnnotationCommentNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type AnnotationLabelNode = Node & {
  __typename?: 'AnnotationLabelNode';
  annotationSet: AnnotationNodeConnection;
  annotationcampaignSet: AnnotationCampaignNodeConnection;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  labelsetSet: LabelSetNodeConnection;
  metadataxLabel?: Maybe<LabelNode>;
  name: Scalars['String']['output'];
  pk: Scalars['PK']['output'];
};
// ...
export type AnnotationNode = Node & {
  __typename?: 'AnnotationNode';
  /** Acoustic features add a better description to the signal */
  acousticFeatures?: Maybe<AcousticFeaturesNode>;
  analysis: SpectrogramAnalysisNode;
  annotationComments: AnnotationCommentNodeConnection;
  annotationPhase: AnnotationPhaseNode;
  annotator?: Maybe<UserNode>;
  /** Expertise level of the annotator. */
  annotatorExpertiseLevel?: Maybe<ApiAnnotationAnnotatorExpertiseLevelChoices>;
  annotatorId: Scalars['PK']['output'];
  confidence?: Maybe<ConfidenceNode>;
  createdAt: Scalars['DateTime']['output'];
  detectorConfiguration?: Maybe<DetectorConfigurationNode>;
  endFrequency?: Maybe<Scalars['Float']['output']>;
  endTime?: Maybe<Scalars['Float']['output']>;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  isUpdateOf?: Maybe<AnnotationNode>;
  isUpdateOfId?: Maybe<Scalars['PK']['output']>;
  label: AnnotationLabelNode;
  lastUpdatedAt: Scalars['DateTime']['output'];
  pk: Scalars['PK']['output'];
  spectrogram: SpectrogramNode;
  startFrequency?: Maybe<Scalars['Float']['output']>;
  startTime?: Maybe<Scalars['Float']['output']>;
  type: AnnotationType;
  updatedTo: AnnotationNodeConnection;
  validations?: Maybe<AnnotationValidationNodeNodeConnection>;
};
// ...
export type AnnotationNodeConnection = {
  __typename?: 'AnnotationNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<AnnotationNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type AnnotationPhaseNode = Node & {
  __typename?: 'AnnotationPhaseNode';
  annotationCampaign: AnnotationCampaignNode;
  annotationCampaignPk: Scalars['Int']['output'];
  annotationComments: AnnotationCommentNodeConnection;
  annotationFileRanges: AnnotationFileRangeNodeConnection;
  annotationTasks: AnnotationTaskNodeConnection;
  annotations: AnnotationNodeConnection;
  completedTasksCount: Scalars['Int']['output'];
  createdAt: Scalars['DateTime']['output'];
  createdBy: UserNode;
  endedAt?: Maybe<Scalars['DateTime']['output']>;
  endedBy?: Maybe<UserNode>;
  hasAnnotations: Scalars['Boolean']['output'];
  /** The ID of the object */
  id: Scalars['ID']['output'];
  isCompleted: Scalars['Boolean']['output'];
  isOpen: Scalars['Boolean']['output'];
  phase: AnnotationPhaseType;
  pk: Scalars['PK']['output'];
  tasksCount: Scalars['Int']['output'];
  userCompletedTasksCount: Scalars['Int']['output'];
  userTasksCount: Scalars['Int']['output'];
};
// ...
export enum AnnotationType {
  Box = 'Box',
  Point = 'Point',
  Weak = 'Weak'
}
// ...
export type AnnotationValidationNodeNodeConnection = {
  __typename?: 'AnnotationValidationNodeNodeConnection';
  /** Pagination data for this connection. */
  pageInfo: PageInfoExtra;
  /** Contains the nodes in this connection. */
  results: Array<Maybe<AnnotationValidationNode>>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};

/** An enumeration. */
export enum ApiAnnotationAnnotatorExpertiseLevelChoices {
  /** Average */
  A = 'A',
  /** Expert */
  E = 'E',
  /** Novice */
  N = 'N'
}
// ...
export type ConfidenceNode = Node & {
  __typename?: 'ConfidenceNode';
  annotationSet: AnnotationNodeConnection;
  confidenceIndicatorSets: ConfidenceSetNodeConnection;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  isDefault?: Maybe<Scalars['Boolean']['output']>;
  label: Scalars['String']['output'];
  level: Scalars['Int']['output'];
  pk: Scalars['PK']['output'];
};
// ...
export type ConfidenceSetNodeConnection = {
  __typename?: 'ConfidenceSetNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ConfidenceSetNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type DetectorConfigurationNode = Node & {
  __typename?: 'DetectorConfigurationNode';
  annotations: AnnotationNodeConnection;
  configuration?: Maybe<Scalars['String']['output']>;
  detector: DetectorNode;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  pk: Scalars['PK']['output'];
};
// ...
export type LabelNode = Node & {
  __typename?: 'LabelNode';
  acousticDetectors: AcousticDetectorSpecificationNodeConnection;
  /** Other name found in the bibliography for this label */
  associatedNames?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  channelConfigurationDetectorSpecifications: ChannelConfigurationDetectorSpecificationNodeConnection;
  children: LabelNodeConnection;
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  labelSet: AnnotationLabelNodeConnection;
  maxFrequency?: Maybe<Scalars['Int']['output']>;
  meanDuration?: Maybe<Scalars['Float']['output']>;
  minFrequency?: Maybe<Scalars['Int']['output']>;
  nickname?: Maybe<Scalars['String']['output']>;
  parent?: Maybe<LabelNode>;
  plurality?: Maybe<Scalars['SignalPluralityEnum']['output']>;
  relatedBibliography: BibliographyNodeConnection;
  shape?: Maybe<Scalars['SignalShapeEnum']['output']>;
  sound?: Maybe<SoundNode>;
  source: SourceNode;
};
// ...
export type LabelSetNodeConnection = {
  __typename?: 'LabelSetNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<LabelSetNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type Node = {
  /** The ID of the object */
  id: Scalars['ID']['output'];
};
// ...
export type SpectrogramAnalysisNode = Node & {
  __typename?: 'SpectrogramAnalysisNode';
  annotationCampaigns: AnnotationCampaignNodeConnection;
  annotations: AnnotationNodeConnection;
  colormap: ColormapNode;
  createdAt: Scalars['DateTime']['output'];
  /** Duration of the segmented data (in s) */
  dataDuration?: Maybe<Scalars['Float']['output']>;
  dataset: DatasetNode;
  description?: Maybe<Scalars['String']['output']>;
  dynamicMax: Scalars['Float']['output'];
  dynamicMin: Scalars['Float']['output'];
  end?: Maybe<Scalars['DateTime']['output']>;
  endDate?: Maybe<Scalars['DateTime']['output']>;
  fft: FftNode;
  filesCount?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  legacy: Scalars['Boolean']['output'];
  legacyConfiguration?: Maybe<LegacySpectrogramConfigurationNode>;
  name: Scalars['String']['output'];
  owner: UserNode;
  path: Scalars['String']['output'];
  pk: Scalars['PK']['output'];
  spectrograms?: Maybe<SpectrogramNodeNodeConnection>;
  start?: Maybe<Scalars['DateTime']['output']>;
  startDate?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type SpectrogramNode = Node & {
  __typename?: 'SpectrogramNode';
  analysis: SpectrogramAnalysisNodeConnection;
  annotationComments: AnnotationCommentNodeConnection;
  annotationTasks?: Maybe<AnnotationTaskNodeNodeConnection>;
  annotations?: Maybe<AnnotationNodeNodeConnection>;
  audioPath: Scalars['String']['output'];
  duration: Scalars['Int']['output'];
  end: Scalars['DateTime']['output'];
  filename: Scalars['String']['output'];
  format: FileFormatNode;
  /** The ID of the object */
  id: Scalars['ID']['output'];
  path: Scalars['String']['output'];
  pk: Scalars['PK']['output'];
  start: Scalars['DateTime']['output'];
  taskStatus?: Maybe<AnnotationTaskStatus>;
};
// ...
export type UserNode = Node & {
  __typename?: 'UserNode';
  annotationComments: AnnotationCommentNodeConnection;
  annotationFileRanges: AnnotationFileRangeNodeConnection;
  annotationResultsValidation: AnnotationValidationNodeConnection;
  annotationTasks: AnnotationTaskNodeConnection;
  annotationcampaignSet: AnnotationCampaignNodeConnection;
  annotations: AnnotationNodeConnection;
  annotatorGroups: UserGroupNodeConnection;
  archives: ArchiveNodeConnection;
  createdPhases: AnnotationPhaseNodeConnection;
  datasetSet: DatasetNodeConnection;
  dateJoined: Scalars['DateTime']['output'];
  displayName: Scalars['String']['output'];
  email: Scalars['String']['output'];
  endedPhases: AnnotationPhaseNodeConnection;
  expertise?: Maybe<ExpertiseLevel>;
  firstName: Scalars['String']['output'];
  /** The ID of the object */
  id: Scalars['ID']['output'];
  /** Designates whether this user should be treated as active. Unselect this instead of deleting accounts. */
  isActive: Scalars['Boolean']['output'];
  isAdmin: Scalars['Boolean']['output'];
  /** Designates whether the user can log into this admin site. */
  isStaff: Scalars['Boolean']['output'];
  /** Designates that this user has all permissions without explicitly assigning them. */
  isSuperuser: Scalars['Boolean']['output'];
  lastLogin?: Maybe<Scalars['DateTime']['output']>;
  lastName: Scalars['String']['output'];
  pk: Scalars['PK']['output'];
  spectrogramAnalysis: SpectrogramAnalysisNodeConnection;
  /** Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
  username: Scalars['String']['output'];
};